:recycle: refactor: Changed the directories of DTOs
