+++ src/main/java/br/fatec/easycoast/dtos/item/ItemResponse.java
@@ -1,4 +1,6 @@
-package br.fatec.easycoast.dtos;
+package br.fatec.easycoast.dtos.item;
+
+import br.fatec.easycoast.dtos.square.SquareResponse;
 
 public record ItemResponse(
     Integer id,
