+++ src/main/java/br/fatec/easycoast/dtos/item/ItemsOnly.java
@@ -1,4 +1,4 @@
-package br.fatec.easycoast.dtos;
+package br.fatec.easycoast.dtos.item;
 
 public record ItemsOnly(
     Integer id,
