+++ src/main/java/br/fatec/easycoast/dtos/square/SquareItems.java
@@ -1,6 +1,8 @@
-package br.fatec.easycoast.dtos;
+package br.fatec.easycoast.dtos.square;
 
 import java.util.List;
+
+import br.fatec.easycoast.dtos.item.ItemsOnly;
 
 public record SquareItems(
         Integer id,
