+++ src/main/java/br/fatec/easycoast/dtos/square/SquareRequest.java
@@ -1,4 +1,4 @@
-package br.fatec.easycoast.dtos;
+package br.fatec.easycoast.dtos.square;
 
 public record SquareRequest(
     String name
