+++ src/main/java/br/fatec/easycoast/dtos/square/SquareResponse.java
@@ -1,4 +1,4 @@
-package br.fatec.easycoast.dtos;
+package br.fatec.easycoast.dtos.square;
 
 public record SquareResponse(
     Integer id,
