+++ src/main/java/br/fatec/easycoast/dtos/user/UserProfile.java
@@ -1,4 +1,4 @@
-package br.fatec.easycoast.dtos;
+package br.fatec.easycoast.dtos.user;
 
 public enum UserProfile {
     ADMIN,
