+++ src/main/java/br/fatec/easycoast/mappers/SquareMapper.java
@@ -1,8 +1,8 @@
 package br.fatec.easycoast.mappers;
 
-import br.fatec.easycoast.dtos.SquareItems;
-import br.fatec.easycoast.dtos.SquareRequest;
-import br.fatec.easycoast.dtos.SquareResponse;
+import br.fatec.easycoast.dtos.square.SquareItems;
+import br.fatec.easycoast.dtos.square.SquareRequest;
+import br.fatec.easycoast.dtos.square.SquareResponse;
 import br.fatec.easycoast.entities.Square;
 
 public class SquareMapper {
