+++ src/main/java/br/fatec/easycoast/mappers/UserMapper.java
@@ -1,7 +1,7 @@
 package br.fatec.easycoast.mappers;
 
-import br.fatec.easycoast.dtos.UserRequest;
-import br.fatec.easycoast.dtos.UserResponse;
+import br.fatec.easycoast.dtos.user.UserRequest;
+import br.fatec.easycoast.dtos.user.UserResponse;
 import br.fatec.easycoast.entities.User;
 
 public class UserMapper {
