+++ src/main/java/br/fatec/easycoast/services/ItemService.java
@@ -6,8 +6,8 @@ import java.util.stream.Collectors;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import br.fatec.easycoast.dtos.ItemRequest;
-import br.fatec.easycoast.dtos.ItemResponse;
+import br.fatec.easycoast.dtos.item.ItemRequest;
+import br.fatec.easycoast.dtos.item.ItemResponse;
 import br.fatec.easycoast.entities.Item;
 import br.fatec.easycoast.mappers.ItemMapper;
 import br.fatec.easycoast.repositories.ItemRepository;
