+++ src/main/java/br/fatec/easycoast/services/UserService.java
@@ -2,8 +2,9 @@ package br.fatec.easycoast.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import br.fatec.easycoast.dtos.UserRequest;
-import br.fatec.easycoast.dtos.UserResponse;
+
+import br.fatec.easycoast.dtos.user.UserRequest;
+import br.fatec.easycoast.dtos.user.UserResponse;
 import br.fatec.easycoast.entities.User;
 import br.fatec.easycoast.mappers.UserMapper;
 import br.fatec.easycoast.repositories.UserRepository;
